requirements.txt
python-binance
pandas
TA-Lib
scikit-learn
numpy
PyQt5